Move compute_team_row from _health_ranking.py to snapshot.py
- _health_ranking no longer imports compute_raw from _dev_repo_metrics — dependency is now unidirectional
- snapshot imports compute_raw and median directly
- _health_ranking is now pure mapping: band_from_health, raw_to_row, rank_rows
